+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-## [1.0.0] - TODO: Add release date.
-
-* TODO: Describe initial release.
+## [1.0.0] - First publish frame_animate_widget.
+## [1.1.0] - Add reset function to animate.